==> notebooks/01_Sampling_AIS.ipynb <==
# this cell contains the code to access GitLab repo
# need it to install ais package from GitLab repo
import sys
import subprocess

GITLAB_USER = "read_aistt"  # read only access
GITLAB_TOKEN = "MMQ6ky1rnLsuKxjyZuvB"

# clone the repo and install the ais packag
git_package = f"git+https://{GITLAB_USER}:{GITLAB_TOKEN}@code.officialstatistics.org/trade-task-team-phase-1/ais.git"

std_out = subprocess.run([sys.executable, "-m", "pip", "install", git_package], capture_output=True, text=True).stdout
print(std_out) 

#allow multiple outputs in one jupyter cell
from IPython.core.interactiveshell import InteractiveShell 
InteractiveShell.ast_node_interactivity = "all"


import pandas as pd
from datetime import datetime
# to apply aggregation functions on spark df
import pyspark.sql.functions as F

# import get_ais() from ais package
from ais import functions as af

# Example 1: Retrieve data for a single date using get_ais()
columns = [
    "mmsi", 
    "latitude", 
    "longitude", 
    'vessel_name',
    "vessel_type", 
    'vessel_type_main',
    'vessel_type_sub',
    'vessel_type_code',
    'vessel_type_cargo',
    'vessel_class',
    "dt_insert_utc", 
    "eeid",
    'length','width','flag_country','destination','eta','source',
    "H3_int_index_1",
    "H3_int_index_5"
]

# date inputs should be in date time format. dt_insert_utc is the basis for the parquet partitions
start_date = datetime.fromisoformat("2022-01-01")
end_date = datetime.fromisoformat("2022-06-30")

# spark is the current spark session you are using. Gets automatically created during kernel init. 
df = af.get_ais(
    spark, 
    start_date = start_date,
    end_date = end_date,
    columns = columns
)

sample_mmsi = '311922000'
SAVEPATH = "s3a://team-cedarsdata-582958291898-mvvpu/"
SEED = 2022
h3_id = [581118283558682623, 581109487465660415, 581122681605193727, 581135875744727039, 581267817140060159, 581514107744681983, 58151850579119308, 581760398349303807, 581500913605148671, 581509709698170879, 581742806163259391, 581764796395814911, 581769194442326015, 582081455744614399]

len(h3_id)

# columns in df
df.columns

# name and type of each column in df
df.printSchema()

# display data contained in 1st row of df. 
# Each row represents a single AIS message transmitted by a single ship
df.show(n=1, vertical=True, truncate=False)

df = df.filter(F.col('H3_int_index_1').isin(h3_id)).persist()
df.show()

df.count()

df.select(F.countDistinct("mmsi")).show()

#filter for sample mmsi
# df_sample = df.filter(F.col('mmsi')==sample_mmsi)
df_sample = df.sample(0.05, seed=SEED).persist()
df_sample

df_sample.count()

unique_mmsi = df_sample.dropDuplicates(['mmsi']).select("mmsi").toPandas()
unique_mmsi

print("coverage mmsi from sample: {:.4}%".format(65901/72891*100))

# write as parquet to s3
df_sample.repartition('mmsi').write.mode('overwrite').parquet(SAVEPATH+"AIS_sample_baltic_005_202201_202206a.parquet")

# write as parquet to local
df_sample.repartition('mmsi').write.mode('overwrite').parquet("AIS_sample_baltic_005_202201_202206a.parquet")

# write as parquet to our own s3
df_sample.limit(1000).repartition('mmsi').write.mode('overwrite').parquet("s3://know-ais/data/AIS_sample_baltic_005_202201_202206.parquet")

!ls

!pip install -q boto3

import boto3
s3 = boto3.resource('s3')
s3.meta.client.upload_file('AIS_sample_baltic_025_202201_202206.parquet', 'knowais-unbdh', 'knowais/AIS_sample_baltic_025_202201_202206.parquet')
