==> notebooks/04_Graph_Analyisis_Port.ipynb <==
from graphdatascience import GraphDataScience
import pandas as pd

# Use Neo4j URI and credentials according to your setup
# NEO4J_URI could look similar to "bolt://my-server.neo4j.io:7687"
ip="localhost"
port="7687"
pwd="bps12345"
gds = GraphDataScience("bolt://"+ip+":"+port, auth=("neo4j", pwd))

# Check the installed GDS version on the server
print(gds.version())
assert gds.version()

# set database
gds.set_database('neo4j')

# list gds function
gds.list()

# Create a port relation graph projection
gds.run_cypher(
  """
  CALL gds.graph.project(
    "port-relation",
    "Port",
    {Visit: {orientation: 'UNDIRECTED', properties:'count'}}
  )
  """
)

#graph project for path finding analysis
gds.run_cypher(
  """
  CALL gds.graph.project(
    "port-path",
    [
      "Port","Shipping"
    ],
    [
      {Origin: {orientation: 'UNDIRECTED', properties:'distance'}},
      {Destination: {orientation: 'UNDIRECTED', properties:'distance'}}
    ]
  )
  """
)

# Project the graph into the GDS Graph Catalog
G_port_relation = gds.graph.get('port-relation')
G_port_relation.node_count()

# Project the graph into the GDS Graph Catalog
G_port_path = gds.graph.get('port-path')
G_port_path.node_count()

## find shortest path using yens algorithm
source_id = gds.find_node_id(["Port"], {"port_name": "Agio Theodoroi"})
target_id = gds.find_node_id(["Port"], {"port_name": "YSTAD"})

res = gds.shortestPath.yens.stream(G_port_path, sourceNode=source_id, targetNode=target_id, k =5, relationshipWeightProperty="distance")
res

# community detection
louvain_result = gds.run_cypher(
    """
    CALL gds.louvain.stream('port-relation')
    YIELD nodeId, communityId
    RETURN 
        gds.util.asNode(nodeId).port_id AS id, 
        gds.util.asNode(nodeId).port_name AS port,
        gds.util.asNode(nodeId).country_code AS country,
        gds.util.asNode(nodeId).area AS area, 
        communityId
    ORDER BY communityId ASC
    """
)
louvain_result.to_csv('result/port_community_detection_louvain.csv',index=False)
louvain_result

louvain_result[louvain_result['communityId']==38]

louvain_result['communityId'].value_counts()

# port importance
pagerank_result = gds.run_cypher(
    """
    CALL gds.pageRank.stream('port-relation',{
        maxIterations: 100,
        dampingFactor: 0.85,
        relationshipWeightProperty: 'count'
    })
    YIELD nodeId, score
    RETURN 
        gds.util.asNode(nodeId).port_id AS id, 
        gds.util.asNode(nodeId).port_name AS port, 
        gds.util.asNode(nodeId).country_code AS country, 
        score
    ORDER BY score DESC
    """
)
pagerank_result.to_csv("result/port_centrality_pagerank.csv",index=False)
pagerank_result

# country importance using betweenness
betweenness_result = gds.run_cypher(
    """
    CALL gds.betweenness.stream('port-relation',{
        relationshipWeightProperty: 'count'
    })
    YIELD nodeId, score
    RETURN 
        gds.util.asNode(nodeId).port_id AS id, 
        gds.util.asNode(nodeId).port_name AS port, 
        score
    ORDER BY score DESC
    """
)
betweenness_result.to_csv("result/port_centrality_betweenness.csv",index=False)
betweenness_result

# The CELF algorithm for influence maximization aims to find k nodes that maximize the expected spread of influence in the network. 
# Leskovec et al. 2007 introduced the CELF algorithm in their study Cost-effective Outbreak Detection in Networks
celf_result = gds.run_cypher(
    """
    CALL gds.beta.influenceMaximization.celf.stream('port-relation',{
        seedSetSize: 261
    })
    YIELD nodeId, spread
    RETURN 
        gds.util.asNode(nodeId).port_id AS id, 
        gds.util.asNode(nodeId).port_name AS port,
        spread
    ORDER BY spread DESC
    """
)
celf_result.to_csv('result/port_centrality_celf.csv',index=False)
celf_result

embedding_res = gds.run_cypher(
    """
    CALL gds.fastRP.stream('port-relation',
    {
        embeddingDimension: 10,
        randomSeed: 2022,
        relationshipWeightProperty: 'count'
    }
    )
    YIELD nodeId, embedding
    RETURN 
        gds.util.asNode(nodeId).port_id AS id, 
        gds.util.asNode(nodeId).port_name AS port,
        embedding
    """
)
embedding_res.to_feather('result/port_embedding_fastrp.f')
embedding_res
